Remove fork utils from stfn
